🔨 [#282] - Extract CashAdjustments field-casting and UOM-conversion duplication 🧹
- 🔨 Added CastsRequestFields trait (castToFloat/castToBoolean) and applied it to 10 CashAdjustments FormRequests, replacing repeated string-to-native prepareForValidation() casts
- 🔨 Added ConvertsUomQuantities trait (getConversion/convertToBaseQuantity) shared by StockOutService and OpeningBalanceService, removing a byte-identical duplicated method
- 📝 Created task file for issue #282 tracking the full SonarCloud duplication cleanup (9 clusters, 2 complete)
