+++ code/api/app/Http/Requests/CashAdjustments/CashTerminals/UpdateCashTerminalRequest.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Requests\CashAdjustments\CashTerminals;
 
+use App\Http\Requests\Concerns\CastsRequestFields;
 use Illuminate\Foundation\Http\FormRequest;
 
 /**
@@ -18,6 +19,8 @@ use Illuminate\Foundation\Http\FormRequest;
  */
 class UpdateCashTerminalRequest extends FormRequest
 {
+    use CastsRequestFields;
+
     public function authorize(): bool
     {
         return $this->user()->can('update', $this->route('cashTerminal'));
@@ -45,10 +48,6 @@ class UpdateCashTerminalRequest extends FormRequest
 
     public function prepareForValidation(): void
     {
-        if ($this->has('is_active') && is_string($this->is_active)) {
-            $this->merge([
-                'is_active' => filter_var($this->is_active, FILTER_VALIDATE_BOOLEAN),
-            ]);
-        }
+        $this->castToBoolean('is_active');
     }
 }
